# docs/api-swagger.yaml
openapi: 3.0.3
info:
  title: Go Data Sync HTTP API
  description: |
    # Real-Time MongoDB Data Synchronization API

    This API provides comprehensive endpoints for managing real-time MongoDB data synchronization between cloud and VM environments. The system supports OAuth2 authentication, adaptive performance optimization, and real-time monitoring.

    ## Key Features
    - **OAuth2 Client Credentials Flow** for secure authentication
    - **Real-time Data Synchronization** using MongoDB change streams
    - **Adaptive Performance Control** with automatic optimization
    - **Comprehensive Monitoring** with metrics and health endpoints
    - **WebSocket Support** for real-time updates
    - **TCP and HTTP Transport** protocols

    ## Authentication
    Most endpoints require OAuth2 Bearer token authentication. Use the `/api/auth/token` endpoint to obtain access tokens.

    ## Rate Limiting
    - Global: 100 requests/second (burst: 200)
    - Per-client: 20 requests/second (burst: 40)
    - Per-IP connections: 10/minute

  version: "2.0.0"
  contact:
    name: Go Data Sync API Support
    url: https://github.com/your-org/go-data-sync-http
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server
security:
  - BearerAuth: []
  - AdminAPIKey: []
paths:
...
  /api/alerts/history:
    get:
      tags:
        - Alerts
      summary: Get Alert History
      description: Retrieve historical alert data with pagination
      operationId: getAlertHistory
      parameters:
        - name: limit
          in: query
          description: Number of alerts to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Number of alerts to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Alert history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertHistoryResponse'
  /api/alerts/rules:
    get:
      tags:
        - Alerts
      summary: Get Alert Rules
      description: Retrieve configured alert rules
      operationId: getAlertRules
      responses:
        '200':
          description: Alert rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRulesResponse'
    post:
      tags:
        - Alerts
      summary: Create Alert Rule
      description: Create a new alert rule
      operationId: createAlertRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertRuleRequest'
      responses:
        '201':
          description: Alert rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
  /api/alerts/rules/{id}:
    put:
      tags:
        - Alerts
      summary: Update Alert Rule
      description: Update an existing alert rule
      operationId: updateAlertRule
      parameters:
        - name: id
          in: path
          required: true
          description: Alert rule ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlertRuleRequest'
      responses:
        '200':
          description: Alert rule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
    delete:
      tags:
        - Alerts
      summary: Delete Alert Rule
      description: Delete an alert rule
      operationId: deleteAlertRule
      parameters:
        - name: id
          in: path
          required: true
          description: Alert rule ID
          schema:
            type: string
      responses:
        '204':
          description: Alert rule deleted

  # ==================== BUFFER-FREE SYSTEM ENDPOINTS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OAuth2 JWT Bearer token authentication. Obtain tokens via `/api/auth/token` endpoint.
        
        **Usage:** `Authorization: Bearer <token>`
    
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: |
        Admin API key for administrative operations. Set via environment variable.
        
        **Usage:** `X-Admin-Key: <admin_key>`
    
    LicenseAuth:
      type: apiKey
      in: header
      name: X-License-Key
      description: |
        License key authentication for VM-sync clients.
        
        **Usage:** `X-License-Key: <license_key>`
...